+++ spec/unit/client/active_record/apns/app_spec.rb
@@ -24,6 +24,6 @@ describe Rpush::Client::ActiveRecord::App do
     app = Rpush::Client::ActiveRecord::Apns::App.new name: 'test', environment: 'development',
                                                      certificate: TEST_CERT_WITH_PASSWORD, password: 'incorrect'
     app.valid?
-    app.errors[:certificate].should eq ["value must contain a certificate and a private key."]
+    app.errors[:certificate].should eq ['value must contain a certificate and a private key.']
   end
 end